Remove some oddly-slow lambdas on hot paths; use specialized replacements
Comparators.comparing() for float values results in some sorta-slow code, calling Float.floatToIntBits() and Float.intBitsToFloat() several times when it can be reduced to one call to Float.floatToIntBits() and some subtraction. This version won't really handle NaN well, but if your graph has NaN for a distance, it isn't a usable graph. It also treats -0.0f as less than 0.0f, but I doubt this is an issue. This version is about 15% faster, give or take 5%.