Wondering why SquidLib's version with almost the same code was slower...
It turns out the BinaryHeap implementation used here has an uncommon logic bug that was fixed in SquidLib and libGDX 1.9.11 but not in this 1.9.10-based BinaryHeap. The test that I wrote and Nathan Sweet polished up for libGDX is now here to test on this BinaryHeap; currently the test fails.