Add a big  test that helped find a tiny fix for a crash
The checkLength() method was called after a node was hashed, but before it was added (at the old hash, in the new larger table). This would cause problems when the table resized during the insertion of a node that would have its hash change.

Feel free to change the test to something that doesn't have a test dependency on SquidLib; this dep doesn't affect users of simple-graphs, but using JitPack stops simple-graphs from being uploaded to Maven Central.